+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Phase 2 — full apply to clean up old S3 objects and apply remaining changes:
 
 ### Added
 
+- **CUMULUS-4710**
+  - Implement list of stats route in iceberg search api
 - **CUMULUS-4708**
   - Implement list of executions route in iceberg search api
 - **CUMULUS-4707**